# butter/public/swagger.json
{
	"definitions": {
		"CreateMemerRequest": {
			"description": "",
			"properties": {
				"username": {
					"type": "string",
					"description": ""
				},
				"password": {
					"type": "string",
					"description": ""
				},
				"displayName": {
					"type": "string",
					"description": ""
				},
				"discordUserId": {
					"type": "string",
					"description": ""
				},
				"profilePic": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"username",
				"password"
			]
		},
		"CreateMemeRequest": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"data": {
					"type": "string",
					"description": ""
				},
				"updoots": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"downdoots": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"creatorId": {
					"type": "object",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"data",
				"creatorId"
			]
		},
		"PatchMemeRequest": {
			"description": "",
			"properties": {
				"name": {
					"type": "string",
					"description": ""
				},
				"data": {
					"type": "string",
					"description": ""
				},
				"updoots": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"downdoots": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object"
		},
		"PatchMemerRequest": {
			"description": "",
			"properties": {
				"username": {
					"type": "string",
					"description": ""
				},
				"displayName": {
					"type": "string",
					"description": ""
				},
				"discordUserId": {
					"type": "string",
					"description": ""
				},
				"profilePic": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object"
		}
	},
	"info": {
		"description": "Backend service for the Oatmeme project",
		"license": {
			"name": "GPL-3.0"
		},
		"title": "butter",
		"version": "0.0.4-dev.0"
	},
	"paths": {
		"/login": {
			"post": {
				"operationId": "Auth_login",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"Auth"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"username": {
									"type": "string",
									"description": ""
								},
								"password": {
									"type": "string",
									"description": ""
								}
							}
						}
					}
				]
			}
		},
		"/register": {
			"post": {
				"operationId": "Auth_register",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"Auth"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateMemerRequest"
						}
					}
				]
			}
		},
		"/memes/{id}": {
			"get": {
				"operationId": "Meme_get",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memes"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Meme_update",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memes"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"description": "",
						"in": "body",
						"name": "meme",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PatchMemeRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Meme_delete",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memes"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "memeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memes": {
			"get": {
				"operationId": "Meme_list",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memes"
				],
				"parameters": []
			},
			"post": {
				"operationId": "Meme_create",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memes"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "meme",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateMemeRequest"
						}
					}
				]
			}
		},
		"/memers/{id}": {
			"get": {
				"operationId": "Memer_get",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memers"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Memer_update",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memers"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"description": "",
						"in": "body",
						"name": "memer",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PatchMemerRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Memer_delete",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memers"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "memerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/memers": {
			"get": {
				"operationId": "Memer_list",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memers"
				],
				"parameters": []
			},
			"post": {
				"operationId": "Memer_create",
				"produces": [
					"text/html"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "",
				"tags": [
					"memers"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "memer",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateMemerRequest"
						}
					}
				]
			}
		}
	},
	"swagger": "2.0",
	"securityDefinitions": {}
}